move continue token
